chore: refactored project driver: Jonas
