+++ src/go.mod
@@ -1,0 +1,7 @@
+module github.com/Agile-Arch-Angels/decentralized-tech-radar_dev
+
+go 1.21.0
+
+require (
+	github.com/gocarina/gocsv v0.0.0-20231116093920-b87c2d0e983a
+)
